server/requirements.txt
psycopg2-binary
pytest==8.1.1
Flask==3.0.3
Flask-Cors==4.0.1
ruff==0.5.6
waitress==3.0.0

